refactor(oni-mcp): split farming tool responsibilities
