+++ mods/oni_mcp/Tools/Impl/Bio/FarmingTools.cs
@@ -9,7 +9,7 @@ using OniMcp.Support;
 
 namespace OniMcp.Tools
 {
-    public static class FarmingTools
+    public static partial class FarmingTools
     {
         private const int CancelEvent = 2127324410;
 
@@ -228,703 +228,5 @@ namespace OniMcp.Tools
             };
         }
 
-        public static McpTool ListSeedCatalog()
-        {
-            return new McpTool
-            {
-                Name = "farming_seed_catalog",
-                Group = "farming",
-                Mode = "read",
-                Risk = "none",
-                Hidden = true,
-                Aliases = new List<string> { "plantable_seeds_list", "seeds_catalog" },
-                Tags = new List<string> { "farming", "plants", "seeds", "catalog" },
-                Description = "兼容入口：请优先使用 farming_planting_control action=seed_catalog。列出 PlantableSeed prefab 及其标签事实；planterBoxCandidate 仅表示 CropSeed 候选，最终兼容性必须由具体 plot 的 IsValidEntity/预览状态验证。",
-                Parameters = new Dictionary<string, McpToolParameter>
-                {
-                    ["query"] = new McpToolParameter { Type = "string", Description = "按种子、植物或显示名筛选", Required = false },
-                    ["limit"] = new McpToolParameter { Type = "integer", Description = "最多返回数量，默认 100，最大 500", Required = false },
-                    ["cropSeedOnly"] = new McpToolParameter { Type = "boolean", Description = "仅返回带 CropSeed 标签的候选种子，默认 false；不代表与任意具体种植槽兼容", Required = false }
-                },
-                Handler = args =>
-                {
-                    string query = args["query"]?.ToString();
-                    int limit = Math.Max(1, Math.Min(ToolUtil.GetInt(args, "limit") ?? 100, 500));
-                    bool cropSeedOnly = ToolUtil.GetBool(args, "cropSeedOnly", false);
-                    var seeds = Assets.Prefabs
-                        .Where(prefab => prefab != null && prefab.gameObject != null)
-                        .Select(prefab => prefab.GetComponent<PlantableSeed>())
-                        .Where(seed => seed != null)
-                        .Where(seed => SeedMatches(seed, query))
-                        .Where(seed => !cropSeedOnly || HasCropSeedTag(seed))
-                        .OrderBy(seed => TargetName(seed.gameObject))
-                        .Take(limit)
-                        .Select(SeedInfo)
-                        .ToList();
-
-                    return CallToolResult.Text(JsonConvert.SerializeObject(new Dictionary<string, object>
-                    {
-                        ["returned"] = seeds.Count,
-                        ["cropSeedOnly"] = cropSeedOnly,
-                        ["hint"] = "cropSeedOnly/planterBoxCandidate only report the CropSeed tag. Validate the chosen seed against the concrete plot with set_planting diagnostics or a plot preview.",
-                        ["seeds"] = seeds
-                    }, McpJsonUtil.Settings));
-                }
-            };
-        }
-
-        public static McpTool SetPlanting()
-        {
-            return new McpTool
-            {
-                Name = "farming_planting_set",
-                Group = "farming",
-                Mode = "write",
-                Risk = "medium",
-                Hidden = true,
-                Aliases = new List<string> { "planter_set_seed", "plant_seed_select" },
-                Tags = new List<string> { "farming", "plants", "planters", "seeds" },
-                Description = "兼容入口：请优先使用 colony_control domain=bio bioDomain=farming action=set。设置或取消种植箱/农砖的种子请求；可选 removeOccupant=true 先铲除当前植物",
-                Parameters = LookupParams(new Dictionary<string, McpToolParameter>
-                {
-                    ["seedTag"] = new McpToolParameter { Type = "string", Description = "种子 prefab/tag，例如 BasicPlantSeed；action=set 时必填", Required = false },
-                    ["mutationTag"] = new McpToolParameter { Type = "string", Description = "植物突变/亚种 tag；未使用突变时留空", Required = false },
-                    ["action"] = new McpToolParameter { Type = "string", Description = "set 或 cancel，默认 set", Required = false, EnumValues = new List<string> { "set", "cancel" } },
-                    ["removeOccupant"] = new McpToolParameter { Type = "boolean", Description = "若已有植物，是否先调用 OrderRemoveOccupant 铲除，默认 false", Required = false },
-                    ["confirm"] = new McpToolParameter { Type = "boolean", Description = "removeOccupant=true 时必须为 true", Required = false }
-                }),
-                Handler = args =>
-                {
-                    var plot = FindPlot(args);
-                    if (plot == null)
-                        return CallToolResult.Error("PlantablePlot target not found");
-
-                    string action = (args["action"]?.ToString() ?? "set").Trim().ToLowerInvariant();
-                    if (action == "cancel")
-                    {
-                        plot.CancelActiveRequest();
-                        return CallToolResult.Text(JsonConvert.SerializeObject(PlotInfo(plot), McpJsonUtil.Settings));
-                    }
-
-                    bool removeOccupant = ToolUtil.GetBool(args, "removeOccupant", false);
-                    if (plot.Occupant != null)
-                    {
-                        if (!removeOccupant)
-                            return CallToolResult.Error("Plot already has an occupant; pass removeOccupant=true and confirm=true to uproot first");
-                        if (!ToolUtil.GetBool(args, "confirm", false))
-                            return CallToolResult.Error("confirm=true is required when removeOccupant=true");
-                        plot.OrderRemoveOccupant();
-                    }
-
-                    string seedName = args["seedTag"]?.ToString();
-                    if (string.IsNullOrWhiteSpace(seedName))
-                        return CallToolResult.Error("seedTag is required for action=set");
-                    var seedTag = TagManager.Create(seedName.Trim());
-                    var seedPrefab = Assets.GetPrefab(seedTag);
-                    var seed = seedPrefab == null ? null : seedPrefab.GetComponent<PlantableSeed>();
-                    if (seed == null)
-                        return CallToolResult.Error("seedTag is not a PlantableSeed prefab");
-                    // Live bug: PlanterBox.possibleDepositObjectTags is often [CropSeed], while
-                    // HasDepositTag(BasicSingleHarvestPlantSeed) checks the seed prefab tag itself
-                    // and returns false even when the seed carries CropSeed and IsValidEntity is true.
-                    bool hasDeposit = SeedMatchesPlotDepositTags(plot, seedPrefab, seedTag);
-                    bool validEntity = plot.IsValidEntity(seedPrefab);
-                    if (!hasDeposit || !validEntity)
-                    {
-                        var plotInfo = PlotInfo(plot);
-                        var seedInfo = SeedInfo(seed);
-                        return CallToolResult.Error(JsonConvert.SerializeObject(new Dictionary<string, object>
-                        {
-                            ["error"] = "Seed is not valid for this planter direction/type",
-                            ["hasDepositTag"] = hasDeposit,
-                            ["isValidEntity"] = validEntity,
-                            ["seed"] = seedInfo,
-                            ["plot"] = plotInfo,
-                            ["hint"] = "Use seed_catalog (cropSeedOnly=true). Deposit checks use seed tags (e.g. CropSeed), not only the seed prefab id."
-                        }, McpJsonUtil.Settings));
-                    }
-
-                    var mutationTag = string.IsNullOrWhiteSpace(args["mutationTag"]?.ToString())
-                        ? Tag.Invalid
-                        : TagManager.Create(args["mutationTag"].ToString().Trim());
-                    float worldInventoryAmount = AvailableSeedAmount(plot.gameObject, seedTag);
-                    plot.CancelActiveRequest();
-                    plot.CreateOrder(seedTag, mutationTag);
-                    var after = PlotInfo(plot);
-                    bool requestedPlantingMatches = RequestedPlantingMatches(plot, seedTag, mutationTag);
-                    bool requestActive = plot.GetActiveRequest != null;
-                    if (!requestedPlantingMatches)
-                    {
-                        return CallToolResult.Error(JsonConvert.SerializeObject(new Dictionary<string, object>
-                        {
-                            ["error"] = "CreateOrder did not establish the requested seed and active planting request",
-                            ["seedTag"] = seedTag.Name,
-                            ["mutationTag"] = mutationTag.IsValid ? mutationTag.Name : null,
-                            ["requestedPlantingMatches"] = requestedPlantingMatches,
-                            ["hasActiveRequest"] = requestActive,
-                            ["hasDepositTag"] = hasDeposit,
-                            ["isValidEntity"] = validEntity,
-                            ["worldInventoryAmount"] = Math.Round(worldInventoryAmount, 3),
-                            ["inventoryNote"] = "World inventory amount is informational only; it does not prove fetchability or explain CreateOrder outcome.",
-                            ["seed"] = SeedInfo(seed),
-                            ["plot"] = after,
-                            ["hint"] = "Inspect plot.validPreview, requestedSeed, hasActiveRequest, occupant, and acceptedSeedTags; validate the exact seed against this plot before retrying."
-                        }, McpJsonUtil.Settings));
-                    }
-
-                    return CallToolResult.Text(JsonConvert.SerializeObject(after, McpJsonUtil.Settings));
-                }
-            };
-        }
-
-        public static McpTool BatchSetPlanting()
-        {
-            return new McpTool
-            {
-                Name = "farming_planting_batch_set",
-                Group = "farming",
-                Mode = "write",
-                Risk = "medium",
-                Hidden = true,
-                Aliases = new List<string> { "planters_batch_set_seed", "plant_seed_area_set" },
-                Tags = new List<string> { "farming", "plants", "planters", "seeds", "batch" },
-                Description = "兼容入口：请优先使用 colony_control domain=bio bioDomain=farming action=batch。按区域批量设置或取消种植请求；适合一次给多块农砖/种植箱安排同一种种子",
-                Parameters = RectParams(new Dictionary<string, McpToolParameter>
-                {
-                    ["seedTag"] = new McpToolParameter { Type = "string", Description = "种子 prefab/tag，例如 BasicPlantSeed；action=set 时必填", Required = false },
-                    ["mutationTag"] = new McpToolParameter { Type = "string", Description = "植物突变/亚种 tag；未使用突变时留空", Required = false },
-                    ["action"] = new McpToolParameter { Type = "string", Description = "set 或 cancel，默认 set", Required = false, EnumValues = new List<string> { "set", "cancel" } },
-                    ["emptyOnly"] = new McpToolParameter { Type = "boolean", Description = "只处理没有当前植物的种植槽，默认 true", Required = false },
-                    ["removeOccupant"] = new McpToolParameter { Type = "boolean", Description = "若已有植物，是否先调用 OrderRemoveOccupant 铲除，默认 false", Required = false },
-                    ["query"] = new McpToolParameter { Type = "string", Description = "按种植槽名称、prefabId、当前植物或请求种子筛选", Required = false },
-                    ["limit"] = new McpToolParameter { Type = "integer", Description = "最多处理数量，默认 100，最大 500", Required = false },
-                    ["confirm"] = new McpToolParameter { Type = "boolean", Description = "必须为 true；区域超过 100 格或 removeOccupant=true 时也必须为 true", Required = true }
-                }),
-                Handler = args =>
-                {
-                    if (!HasRectInput(args))
-                        return CallToolResult.Error("areaId or x1/y1/x2/y2 are required");
-                    if (!ToolUtil.GetBool(args, "confirm", false))
-                        return CallToolResult.Error("confirm=true is required to batch change planting requests");
-
-                    var rect = ToolUtil.GetRect(args);
-                    int cells = RectCellCount(rect);
-                    if (cells > 100 && !ToolUtil.GetBool(args, "confirm", false))
-                        return CallToolResult.Error("confirm=true is required when changing planting requests in more than 100 cells");
-
-                    string action = (args["action"]?.ToString() ?? "set").Trim().ToLowerInvariant();
-                    bool isSet = action == "set";
-                    if (!isSet && action != "cancel")
-                        return CallToolResult.Error("action must be set or cancel");
-
-                    Tag seedTag = Tag.Invalid;
-                    GameObject seedPrefab = null;
-                    if (isSet)
-                    {
-                        string seedName = args["seedTag"]?.ToString();
-                        if (string.IsNullOrWhiteSpace(seedName))
-                            return CallToolResult.Error("seedTag is required for action=set");
-                        seedTag = TagManager.Create(seedName.Trim());
-                        seedPrefab = Assets.GetPrefab(seedTag);
-                        if (seedPrefab == null || seedPrefab.GetComponent<PlantableSeed>() == null)
-                            return CallToolResult.Error("seedTag is not a PlantableSeed prefab");
-                    }
-
-                    var mutationTag = string.IsNullOrWhiteSpace(args["mutationTag"]?.ToString())
-                        ? Tag.Invalid
-                        : TagManager.Create(args["mutationTag"].ToString().Trim());
-                    int worldId = ToolUtil.ResolveWorldId(args);
-                    string query = args["query"]?.ToString();
-                    bool emptyOnly = ToolUtil.GetBool(args, "emptyOnly", true);
-                    bool removeOccupant = ToolUtil.GetBool(args, "removeOccupant", false);
-                    int limit = Math.Max(1, Math.Min(ToolUtil.GetInt(args, "limit") ?? 100, 500));
-                    var seedInfo = isSet ? SeedInfo(seedPrefab.GetComponent<PlantableSeed>()) : null;
-                    int changedCount = 0;
-                    int failedCount = 0;
-                    int unchangedCount = 0;
-                    var changed = new List<Dictionary<string, object>>();
-                    var failures = new List<Dictionary<string, object>>();
-                    var unchanged = new List<Dictionary<string, object>>();
-                    foreach (var plot in Components.BuildingCompletes.Items
-                                 .Select(building => building?.GetComponent<PlantablePlot>())
-                                 .Where(plot => plot != null && ToolUtil.GameObjectMatchesWorld(plot.gameObject, worldId))
-                                 .Where(plot => CellInRect(Grid.PosToCell(plot.gameObject), rect, worldId))
-                                 .Where(plot => PlantingMatches(plot, query))
-                                 .Take(limit))
-                    {
-                        if (!isSet)
-                        {
-                            bool hadRequestedTag = plot.requestedEntityTag.IsValid;
-                            bool hadActiveRequest = plot.GetActiveRequest != null;
-                            if (!hadRequestedTag && !hadActiveRequest)
-                            {
-                                unchangedCount++;
-                                AddBatchDetail(unchanged, BatchPlotOutcome("no_active_request", plot));
-                                continue;
-                            }
-
-                            var before = PlotInfo(plot);
-                            plot.CancelActiveRequest();
-                            bool cleared = !plot.requestedEntityTag.IsValid && plot.GetActiveRequest == null;
-                            if (cleared)
-                            {
-                                changedCount++;
-                                AddBatchDetail(changed, PlotInfo(plot));
-                            }
-                            else
-                            {
-                                failedCount++;
-                                var failure = BatchPlotOutcome("cancel_did_not_clear_request", plot);
-                                failure["before"] = before;
-                                AddBatchDetail(failures, failure);
-                            }
-                            continue;
-                        }
-
-                        bool hasDeposit = SeedMatchesPlotDepositTags(plot, seedPrefab, seedTag);
-                        bool validEntity = plot.IsValidEntity(seedPrefab);
-                        if (!hasDeposit || !validEntity)
-                        {
-                            failedCount++;
-                            var failure = BatchPlotOutcome("seed_not_valid_for_plot", plot);
-                            failure["hasDepositTag"] = hasDeposit;
-                            failure["isValidEntity"] = validEntity;
-                            failure["seed"] = seedInfo;
-                            AddBatchDetail(failures, failure);
-                            continue;
-                        }
-                        if (RequestedPlantingMatches(plot, seedTag, mutationTag))
-                        {
-                            unchangedCount++;
-                            AddBatchDetail(unchanged, BatchPlotOutcome("request_already_active", plot));
-                            continue;
-                        }
-                        if (plot.Occupant != null)
-                        {
-                            if (emptyOnly)
-                            {
-                                unchangedCount++;
-                                AddBatchDetail(unchanged, BatchPlotOutcome("occupant_present_empty_only", plot));
-                                continue;
-                            }
-                            if (!removeOccupant)
-                            {
-                                unchangedCount++;
-                                AddBatchDetail(unchanged, BatchPlotOutcome("occupant_present_remove_disabled", plot));
-                                continue;
-                            }
-                            plot.OrderRemoveOccupant();
-                        }
-                        plot.CancelActiveRequest();
-                        plot.CreateOrder(seedTag, mutationTag);
-                        bool requestedPlantingMatches = RequestedPlantingMatches(plot, seedTag, mutationTag);
-                        bool requestActive = plot.GetActiveRequest != null;
-                        if (requestedPlantingMatches)
-                        {
-                            changedCount++;
-                            AddBatchDetail(changed, PlotInfo(plot));
-                        }
-                        else
-                        {
-                            failedCount++;
-                            var failure = BatchPlotOutcome("create_order_postcondition_failed", plot);
-                            failure["requestedPlantingMatches"] = requestedPlantingMatches;
-                            failure["hasActiveRequest"] = requestActive;
-                            failure["hasDepositTag"] = hasDeposit;
-                            failure["isValidEntity"] = validEntity;
-                            failure["worldInventoryAmount"] = Math.Round(AvailableSeedAmount(plot.gameObject, seedTag), 3);
-                            failure["inventoryNote"] = "World inventory amount is informational only; it does not prove fetchability or explain CreateOrder outcome.";
-                            failure["seed"] = seedInfo;
-                            AddBatchDetail(failures, failure);
-                        }
-                    }
-
-                    return CallToolResult.Text(JsonConvert.SerializeObject(new Dictionary<string, object>
-                    {
-                        ["action"] = action,
-                        ["seedTag"] = seedTag.IsValid ? seedTag.Name : null,
-                        ["changed"] = changedCount,
-                        ["failed"] = failedCount,
-                        ["unchanged"] = unchangedCount,
-                        ["worldId"] = worldId,
-                        ["rect"] = rect,
-                        ["detailLimit"] = 20,
-                        ["plots"] = changed,
-                        ["failures"] = failures,
-                        ["unchangedPlots"] = unchanged
-                    }, McpJsonUtil.Settings));
-                }
-            };
-        }
-
-        public static McpTool UprootArea()
-        {
-            return new McpTool
-            {
-                Name = "plants_uproot_area",
-                Group = "farming",
-                Mode = "execute",
-                Risk = "medium",
-                Hidden = true,
-                Aliases = new List<string> { "farming_uproot_area", "plants_cancel_uproot" },
-                Tags = new List<string> { "farming", "plants", "uproot", "harvest" },
-                Description = "兼容入口：请优先使用 colony_control domain=bio bioDomain=farming action=uproot。按区域标记或取消铲除植物",
-                Parameters = RectParams(new Dictionary<string, McpToolParameter>
-                {
-                    ["action"] = new McpToolParameter { Type = "string", Description = "mark 或 cancel，默认 mark", Required = false, EnumValues = new List<string> { "mark", "cancel" } },
-                    ["priority"] = new McpToolParameter { Type = "integer", Description = "铲除差事优先级 1-9，默认 5", Required = false },
-                    ["topPriority"] = new McpToolParameter { Type = "boolean", Description = "是否设为红色最高优先级，默认 false", Required = false },
-                    ["confirm"] = new McpToolParameter { Type = "boolean", Description = "区域超过 100 格时必须为 true；mark 操作建议传 true", Required = false }
-                }),
-                Handler = args =>
-                {
-                    if (!HasRectInput(args))
-                        return CallToolResult.Error("areaId or x1/y1/x2/y2 are required");
-
-                    var rect = ToolUtil.GetRect(args);
-                    int cells = RectCellCount(rect);
-                    if (cells > 100 && !ToolUtil.GetBool(args, "confirm", false))
-                        return CallToolResult.Error("confirm=true is required when changing uproot orders in more than 100 cells");
-
-                    int worldId = ToolUtil.ResolveWorldId(args);
-                    bool mark = (args["action"]?.ToString() ?? "mark").Trim().ToLowerInvariant() != "cancel";
-                    int changed = 0;
-                    var results = new List<Dictionary<string, object>>();
-                    foreach (var uprootable in Components.Uprootables.Items)
-                    {
-                        var go = uprootable?.gameObject;
-                        if (go == null || !ToolUtil.GameObjectMatchesWorld(go, worldId))
-                            continue;
-                        int cell = Grid.PosToCell(go);
-                        if (!CellInRect(cell, rect, worldId))
-                            continue;
-
-                        if (mark)
-                        {
-                            if (!uprootable.CanUproot())
-                                continue;
-                            uprootable.MarkForUproot();
-                            ApplyPriority(go, args);
-                            results.Add(TargetInfo(go, "marked"));
-                        }
-                        else
-                        {
-                            if (!uprootable.IsMarkedForUproot)
-                                continue;
-                            uprootable.ForceCancelUproot();
-                            go.Trigger(CancelEvent);
-                            results.Add(TargetInfo(go, "cancelled"));
-                        }
-                        changed++;
-                    }
-
-                    return CallToolResult.Text(JsonConvert.SerializeObject(new Dictionary<string, object>
-                    {
-                        ["action"] = mark ? "mark" : "cancel",
-                        ["changed"] = changed,
-                        ["worldId"] = worldId,
-                        ["rect"] = rect,
-                        ["targets"] = results.Take(200).ToList(),
-                        ["truncatedTargets"] = Math.Max(0, results.Count - 200)
-                    }, McpJsonUtil.Settings));
-                }
-            };
-        }
-
-        private static Dictionary<string, McpToolParameter> LookupParams(Dictionary<string, McpToolParameter> extra)
-        {
-            var parameters = new Dictionary<string, McpToolParameter>
-            {
-                ["id"] = new McpToolParameter { Type = "integer", Description = "目标对象 InstanceID", Required = false },
-                ["x"] = new McpToolParameter { Type = "integer", Description = "目标格子 X", Required = false },
-                ["y"] = new McpToolParameter { Type = "integer", Description = "目标格子 Y", Required = false },
-                ["worldId"] = new McpToolParameter { Type = "integer", Description = "目标世界 ID；按坐标查找时默认当前激活世界", Required = false }
-            };
-            foreach (var item in extra)
-                parameters[item.Key] = item.Value;
-            return parameters;
-        }
-
-        private static Dictionary<string, McpToolParameter> RectParams(Dictionary<string, McpToolParameter> extra)
-        {
-            var parameters = new Dictionary<string, McpToolParameter>
-            {
-                ["areaId"] = new McpToolParameter { Type = "string", Description = "可选区域句柄；提供后可省略 x1/y1/x2/y2", Required = false },
-                ["x1"] = new McpToolParameter { Type = "integer", Description = "区域左下/起点 X；使用 areaId 时可省略", Required = false },
-                ["y1"] = new McpToolParameter { Type = "integer", Description = "区域左下/起点 Y；使用 areaId 时可省略", Required = false },
-                ["x2"] = new McpToolParameter { Type = "integer", Description = "区域右上/终点 X；使用 areaId 时可省略", Required = false },
-                ["y2"] = new McpToolParameter { Type = "integer", Description = "区域右上/终点 Y；使用 areaId 时可省略", Required = false },
-                ["worldId"] = new McpToolParameter { Type = "integer", Description = "目标世界 ID，默认 areaId 绑定世界或当前激活世界", Required = false }
-            };
-            foreach (var item in extra)
-                parameters[item.Key] = item.Value;
-            return parameters;
-        }
-
-        private static PlantablePlot FindPlot(JObject args)
-        {
-            int? id = ToolUtil.GetInt(args, "id");
-            int? x = ToolUtil.GetInt(args, "x");
-            int? y = ToolUtil.GetInt(args, "y");
-            int? cell = x.HasValue && y.HasValue ? Grid.XYToCell(x.Value, y.Value) : (int?)null;
-            int worldId = cell.HasValue ? ToolUtil.ResolveWorldId(args) : (ToolUtil.GetInt(args, "worldId") ?? -1);
-
-            foreach (var building in Components.BuildingCompletes.Items)
-            {
-                var go = building?.gameObject;
-                var plot = go == null ? null : go.GetComponent<PlantablePlot>();
-                if (plot == null || !ToolUtil.GameObjectMatchesWorld(go, worldId))
-                    continue;
-                var kpid = go.GetComponent<KPrefabID>();
-                if (id.HasValue && kpid != null && kpid.InstanceID == id.Value)
-                    return plot;
-                if (cell.HasValue && Grid.PosToCell(go) == cell.Value)
-                    return plot;
-            }
-            return null;
-        }
-
-        private static HarvestDesignatable FindHarvestable(JObject args)
-        {
-            int? id = ToolUtil.GetInt(args, "id");
-            int? x = ToolUtil.GetInt(args, "x");
-            int? y = ToolUtil.GetInt(args, "y");
-            int? cell = x.HasValue && y.HasValue ? Grid.XYToCell(x.Value, y.Value) : (int?)null;
-            int worldId = cell.HasValue ? ToolUtil.ResolveWorldId(args) : (ToolUtil.GetInt(args, "worldId") ?? -1);
-
-            foreach (var harvestable in Components.HarvestDesignatables.Items)
-            {
-                var go = harvestable?.gameObject;
-                if (go == null || !ToolUtil.GameObjectMatchesWorld(go, worldId))
-                    continue;
-                var kpid = go.GetComponent<KPrefabID>();
-                if (id.HasValue && kpid != null && kpid.InstanceID == id.Value)
-                    return harvestable;
-                if (cell.HasValue && Grid.PosToCell(go) == cell.Value)
-                    return harvestable;
-            }
-            return null;
-        }
-
-        private static Dictionary<string, object> PlotInfo(PlantablePlot plot)
-        {
-            var result = TargetInfo(plot.gameObject, null);
-            result["requestedSeed"] = plot.requestedEntityTag.IsValid ? plot.requestedEntityTag.Name : null;
-            result["requestedMutation"] = plot.requestedEntityAdditionalFilterTag.IsValid ? plot.requestedEntityAdditionalFilterTag.Name : null;
-            result["hasActiveRequest"] = plot.GetActiveRequest != null;
-            result["validPreview"] = plot.ValidPlant;
-            result["acceptsFertilizer"] = plot.AcceptsFertilizer;
-            result["acceptsIrrigation"] = plot.AcceptsIrrigation;
-            result["occupant"] = plot.Occupant == null ? null : TargetInfo(plot.Occupant, null);
-            result["acceptedSeedTags"] = plot.possibleDepositObjectTags.Select(tag => tag.Name).OrderBy(name => name).ToList();
-            return result;
-        }
-
-        private static Dictionary<string, object> HarvestableInfo(HarvestDesignatable harvestable)
-        {
-            var result = TargetInfo(harvestable.gameObject, null);
-            result["canBeHarvested"] = harvestable.CanBeHarvested();
-            result["markedForHarvest"] = harvestable.MarkedForHarvest;
-            result["harvestWhenReady"] = harvestable.HarvestWhenReady;
-            result["inPlanterBox"] = harvestable.InPlanterBox;
-            var harvestableComponent = harvestable.GetComponent<Harvestable>();
-            if (harvestableComponent != null)
-                result["harvestableComponent"] = harvestableComponent.GetType().Name;
-            return result;
-        }
-
-        private static Dictionary<string, object> SeedInfo(PlantableSeed seed)
-        {
-            var kpid = seed.GetComponent<KPrefabID>();
-            var seedGo = seed.gameObject;
-            var seedTag = kpid?.PrefabTag ?? Tag.Invalid;
-            // Tags agents need when matching list_planting.acceptedSeedTags (e.g. CropSeed).
-            var depositTags = new List<string>();
-            if (kpid != null)
-            {
-                foreach (var tag in kpid.Tags)
-                {
-                    if (tag.IsValid && !string.IsNullOrEmpty(tag.Name))
-                        depositTags.Add(tag.Name);
-                }
-                depositTags = depositTags.Distinct().OrderBy(name => name).ToList();
-            }
-
-            // Live play: agents saw acceptedSeedTags=["CropSeed"] but seed_catalog only
-            // returned seedTag/plantId, so they tried BasicSingleHarvestPlantSeed and got
-            // "Seed is not valid for this planter direction/type" with no catalog guidance.
-            return new Dictionary<string, object>
-            {
-                ["seedTag"] = seedTag.IsValid ? seedTag.Name : seedGo.name,
-                ["name"] = TargetName(seedGo),
-                ["plantId"] = seed.PlantID.Name,
-                ["previewId"] = seed.PreviewID.Name,
-                ["direction"] = seed.Direction.ToString(),
-                ["replantGroundTag"] = seed.replantGroundTag.IsValid ? seed.replantGroundTag.Name : null,
-                ["seedTags"] = depositTags,
-                ["hasCropSeedTag"] = depositTags.Any(tag => tag == "CropSeed"),
-                ["isCropSeed"] = depositTags.Any(tag => tag == "CropSeed"),
-                ["planterBoxCandidate"] = depositTags.Any(tag => tag == "CropSeed")
-            };
-        }
-
-        private static bool PlantingMatches(PlantablePlot plot, string query)
-        {
-            if (string.IsNullOrWhiteSpace(query))
-                return true;
-            string q = query.Trim();
-            return Contains(TargetName(plot.gameObject), q)
-                || Contains(plot.gameObject.GetComponent<KPrefabID>()?.PrefabTag.Name ?? plot.gameObject.name, q)
-                || Contains(plot.requestedEntityTag.Name, q)
-                || Contains(plot.Occupant?.GetProperName(), q)
-                || Contains(plot.Occupant?.GetComponent<KPrefabID>()?.PrefabTag.Name, q);
-        }
-
-        private static bool HarvestableMatches(HarvestDesignatable harvestable, string query)
-        {
-            if (string.IsNullOrWhiteSpace(query))
-                return true;
-            string q = query.Trim();
-            var go = harvestable.gameObject;
-            return Contains(TargetName(go), q)
-                || Contains(go.GetComponent<KPrefabID>()?.PrefabTag.Name ?? go.name, q);
-        }
-
-        private static bool SeedMatches(PlantableSeed seed, string query)
-        {
-            if (string.IsNullOrWhiteSpace(query))
-                return true;
-            string q = query.Trim();
-            var kpid = seed.GetComponent<KPrefabID>();
-            return Contains(TargetName(seed.gameObject), q)
-                || Contains(kpid?.PrefabTag.Name ?? seed.gameObject.name, q)
-                || Contains(seed.PlantID.Name, q)
-                || Contains(seed.PreviewID.Name, q)
-                || (kpid != null && kpid.Tags.Any(tag => Contains(tag.Name, q)));
-        }
-
-        private static bool HasCropSeedTag(PlantableSeed seed)
-        {
-            var kpid = seed?.GetComponent<KPrefabID>();
-            return kpid != null && kpid.Tags.Any(tag => tag.Name == "CropSeed");
-        }
-
-        private static bool SeedMatchesPlotDepositTags(PlantablePlot plot, GameObject seedPrefab, Tag seedTag)
-        {
-            if (plot == null || seedPrefab == null)
-                return false;
-            if (plot.HasDepositTag(seedTag))
-                return true;
-            var kpid = seedPrefab.GetComponent<KPrefabID>();
-            if (kpid == null)
-                return false;
-            // Category deposit tags (CropSeed) must match tags on the seed prefab, not only seed prefab id.
-            foreach (var tag in kpid.Tags)
-            {
-                if (tag.IsValid && plot.HasDepositTag(tag))
-                    return true;
-            }
-            return false;
-        }
-
-        private static bool RequestedPlantingMatches(PlantablePlot plot, Tag seedTag, Tag mutationTag)
-        {
-            if (plot == null || plot.GetActiveRequest == null)
-                return false;
-            if (!plot.requestedEntityTag.IsValid || !seedTag.IsValid
-                || !string.Equals(plot.requestedEntityTag.Name, seedTag.Name, StringComparison.OrdinalIgnoreCase))
-            {
-                return false;
-            }
-
-            var requestedMutationTag = plot.requestedEntityAdditionalFilterTag;
-            if (requestedMutationTag.IsValid != mutationTag.IsValid)
-                return false;
-            return !mutationTag.IsValid
-                || string.Equals(requestedMutationTag.Name, mutationTag.Name, StringComparison.OrdinalIgnoreCase);
-        }
-
-        private static Dictionary<string, object> BatchPlotOutcome(string reason, PlantablePlot plot)
-        {
-            return new Dictionary<string, object>
-            {
-                ["reason"] = reason,
-                ["plot"] = PlotInfo(plot)
-            };
-        }
-
-        private static void AddBatchDetail(List<Dictionary<string, object>> details, Dictionary<string, object> item)
-        {
-            if (details.Count < 20)
-                details.Add(item);
-        }
-
-        private static float AvailableSeedAmount(GameObject target, Tag seedTag)
-        {
-            if (target == null || !seedTag.IsValid)
-                return 0f;
-            var world = target.GetMyWorld();
-            if (world == null)
-                return 0f;
-            return ToolUtil.SafeFloat(world.worldInventory.GetTotalAmount(seedTag, includeRelatedWorlds: true));
-        }
-
-        private static void ApplyPriority(GameObject go, JObject args)
-        {
-            var prioritizable = go.GetComponent<Prioritizable>();
-            if (prioritizable == null)
-                return;
-            bool top = ToolUtil.GetBool(args, "topPriority", false);
-            int priority = Math.Max(1, Math.Min(ToolUtil.GetInt(args, "priority") ?? 5, 9));
-            prioritizable.SetMasterPriority(new PrioritySetting(top ? PriorityScreen.PriorityClass.topPriority : PriorityScreen.PriorityClass.basic, top ? 1 : priority));
-        }
-
-        private static Dictionary<string, object> TargetInfo(GameObject go, string status)
-        {
-            int cell = Grid.PosToCell(go);
-            var kpid = go.GetComponent<KPrefabID>();
-            var result = new Dictionary<string, object>
-            {
-                ["id"] = kpid?.InstanceID ?? go.GetInstanceID(),
-                ["name"] = TargetName(go),
-                ["prefabId"] = kpid?.PrefabTag.Name ?? go.name,
-                ["x"] = Grid.IsValidCell(cell) ? Grid.CellColumn(cell) : -1,
-                ["y"] = Grid.IsValidCell(cell) ? Grid.CellRow(cell) : -1,
-                ["worldId"] = Grid.IsValidCell(cell) && Grid.IsWorldValidCell(cell) ? Grid.WorldIdx[cell] : -1
-            };
-            if (!string.IsNullOrEmpty(status))
-                result["status"] = status;
-            return result;
-        }
-
-        private static string TargetName(GameObject go)
-        {
-            return ToolUtil.CleanName(go.GetProperName());
-        }
-
-        private static bool HasRectInput(JObject args)
-        {
-            return !string.IsNullOrWhiteSpace(args["areaId"]?.ToString())
-                || (args["x1"] != null && args["y1"] != null && args["x2"] != null && args["y2"] != null);
-        }
-
-        private static bool CellInRect(int cell, Dictionary<string, int> rect, int worldId)
-        {
-            if (!Grid.IsValidCell(cell)) return false;
-            if (!ToolUtil.CellMatchesWorld(cell, worldId)) return false;
-            int x = Grid.CellColumn(cell);
-            int y = Grid.CellRow(cell);
-            return x >= rect["x1"] && x <= rect["x2"] && y >= rect["y1"] && y <= rect["y2"];
-        }
-
-        private static int RectCellCount(Dictionary<string, int> rect)
-        {
-            return (rect["x2"] - rect["x1"] + 1) * (rect["y2"] - rect["y1"] + 1);
-        }
-
-        private static bool Contains(string value, string query)
-        {
-            return !string.IsNullOrEmpty(value) && value.IndexOf(query, StringComparison.OrdinalIgnoreCase) >= 0;
-        }
     }
 }
